fix: use local timezone midnight for week/month DB filter boundaries
