+++ Pdmt.Api/Services/AnalyticsService.cs
@@ -15,7 +15,8 @@ public class AnalyticsService(AppDbContext db, IConfiguration config) : IAnalyti
 
     public async Task<WeeklySummaryDto> GetWeeklySummaryAsync(Guid userId, DateOnly weekOf)
     {
-        var monday = DateHelper.GetMonday(weekOf);
+        var tz = GetTz();
+        var monday = DateHelper.GetMonday(weekOf, tz);
 
         var events = await db.Events
             .AsNoTracking()
@@ -50,8 +51,6 @@ public class AnalyticsService(AppDbContext db, IConfiguration config) : IAnalyti
             .Take(5)
             .Select(e => new TopEventDto(e.Title, e.Intensity, e.Timestamp))
             .ToList();
-
-        var tz = GetTz();
         var byDayOfWeek = events
             .GroupBy(e => DateHelper.ToLocalDate(e.Timestamp, tz).DayOfWeek)
             .OrderBy(g => ((int)g.Key + 6) % 7)
@@ -110,24 +109,27 @@ public class AnalyticsService(AppDbContext db, IConfiguration config) : IAnalyti
 
     public async Task<CalendarWeekDto> GetCalendarWeekAsync(Guid userId, DateOnly weekOf)
     {
-        var monday = DateHelper.GetMonday(weekOf);
+        var tz = GetTz();
+        var monday = DateHelper.GetMonday(weekOf, tz);
 
         var events = await db.Events
             .AsNoTracking()
             .Include(e => e.EventTags).ThenInclude(et => et.Tag)
             .Where(e => e.UserId == userId && e.Timestamp >= monday && e.Timestamp < monday.AddDays(7))
             .ToListAsync();
-
-        var tz = GetTz();
         var byDay = events.GroupBy(e => DateHelper.ToLocalDate(e.Timestamp, tz)).ToDictionary(g => g.Key, g => g.ToList());
+
+        // monday is UTC — derive local DateOnly for day iteration and DTO dates
+        var mondayDate = DateHelper.ToLocalDate(monday, tz);
 
         var days = new List<CalendarDayDetailsDto>(7);
         for (var i = 0; i < 7; i++)
         {
-            var date = monday.AddDays(i);
-            if (!byDay.TryGetValue(DateHelper.ToLocalDate(date, tz), out var dayEvents))
+            var dayDate = mondayDate.AddDays(i);
+            var dtoDate = new DateTimeOffset(dayDate.ToDateTime(TimeOnly.MinValue, DateTimeKind.Utc), TimeSpan.Zero);
+            if (!byDay.TryGetValue(dayDate, out var dayEvents))
             {
-                days.Add(new CalendarDayDetailsDto(date, 0, 0, 0, 0, 0.0, [], []));
+                days.Add(new CalendarDayDetailsDto(dtoDate, 0, 0, 0, 0, 0.0, [], []));
                 continue;
             }
 
@@ -154,17 +156,21 @@ public class AnalyticsService(AppDbContext db, IConfiguration config) : IAnalyti
                 .Select(g => new TagCountDto(g.Key, g.Count()))
                 .ToList();
 
-            days.Add(new CalendarDayDetailsDto(date, posEvents.Count, negEvents.Count,
+            days.Add(new CalendarDayDetailsDto(dtoDate, posEvents.Count, negEvents.Count,
                 posSum, negSum, dayScore, topPosTags, topNegTags));
         }
 
-        return new CalendarWeekDto(monday, monday.AddDays(6), days);
+        var weekStart = new DateTimeOffset(mondayDate.ToDateTime(TimeOnly.MinValue, DateTimeKind.Utc), TimeSpan.Zero);
+        return new CalendarWeekDto(weekStart, weekStart.AddDays(6), days);
     }
 
     public async Task<CalendarMonthDto> GetCalendarMonthAsync(Guid userId, int year, int month)
     {
-        var from = new DateTimeOffset(year, month, 1, 0, 0, 0, TimeSpan.Zero);
-        var to = from.AddMonths(1);
+        var tz = GetTz();
+        var fromLocal = new DateTime(year, month, 1, 0, 0, 0);
+        var from = new DateTimeOffset(TimeZoneInfo.ConvertTimeToUtc(fromLocal, tz), TimeSpan.Zero);
+        var toLocal = fromLocal.AddMonths(1);
+        var to = new DateTimeOffset(TimeZoneInfo.ConvertTimeToUtc(toLocal, tz), TimeSpan.Zero);
 
         var raw = await db.Events
             .AsNoTracking()
@@ -172,16 +178,16 @@ public class AnalyticsService(AppDbContext db, IConfiguration config) : IAnalyti
             .Select(e => new { e.Timestamp, e.Type, e.Intensity })
             .ToListAsync();
 
-        var tz = GetTz();
         var byDay = raw.GroupBy(e => DateHelper.ToLocalDate(e.Timestamp, tz)).ToDictionary(g => g.Key, g => g.ToList());
 
         var days = new List<CalendarDayLightDto>(DateTime.DaysInMonth(year, month));
         for (var d = 1; d <= DateTime.DaysInMonth(year, month); d++)
         {
-            var date = new DateTimeOffset(year, month, d, 0, 0, 0, TimeSpan.Zero);
-            if (!byDay.TryGetValue(DateHelper.ToLocalDate(date, tz), out var dayEvents))
+            var dayDate = new DateOnly(year, month, d);
+            var dtoDate = new DateTimeOffset(dayDate.ToDateTime(TimeOnly.MinValue, DateTimeKind.Utc), TimeSpan.Zero);
+            if (!byDay.TryGetValue(dayDate, out var dayEvents))
             {
-                days.Add(new CalendarDayLightDto(date, 0, 0, 0.0));
+                days.Add(new CalendarDayLightDto(dtoDate, 0, 0, 0.0));
                 continue;
             }
 
@@ -203,7 +209,7 @@ public class AnalyticsService(AppDbContext db, IConfiguration config) : IAnalyti
             }
             var total = dayEvents.Count;
             var dayScore = total == 0 ? 0.0 : (double)(posSum - negSum) / total;
-            days.Add(new CalendarDayLightDto(date, posCount, negCount, dayScore));
+            days.Add(new CalendarDayLightDto(dtoDate, posCount, negCount, dayScore));
         }
 
         return new CalendarMonthDto(days);
